juridica_model/requirements.txt
# requirements.txt
google-generativeai          # Gemini generativo
langchain-google-genai       # Wrapper de embeddings de Gemini
chromadb                     # Vector store local
google-api-python-client     # Conexión a Drive
google-auth-oauthlib
python-dotenv
PyPDF2
tqdm
gradio